refactor(onboarding and popup): migrate onboarding and popup scripts to TypeScript
- Updated webpack configuration to support TypeScript files for onboarding and popup scripts.
- Converted onboarding/welcome.js and popup.js to TypeScript, enhancing type safety and maintainability.
- Improved logging and error handling in the new TypeScript files.
- Removed deprecated JavaScript files to streamline the codebase.
